Listing and adding expenses for a user (#40)
* feat: started to wok on expense and transaction apis

* feat: some wiring in expenses, more compile errors

* Fixing get expenses return type (VsCode remote)

* feat: started to work on expense_operations

* feat: get_expenses is functional

* feat: create_expense and decimal for money

* feat: invalid id handling and postman collection

* fix: import order for fmt

* fix: removing trailing slashes from endpoints

* fix: review comment fixes

new date constant, more expense dto validation, proper async handling

* fix: more review comments

fixing empty find_expenses, removing useless comments

+++ server/entity/src/lib.rs
@@ -1,3 +1,5 @@
+use sea_orm::prelude::Decimal;
+
 pub mod account_activation;
 pub mod currency_type;
 pub mod expense;
@@ -7,4 +9,4 @@ pub mod transaction;
 pub mod user;
 
 pub type Id = i64;
-pub type MoneyAmount = i64;
+pub type MoneyAmount = Decimal;
